backend/requirements.txt
fastapi
uvicorn[standard]
databases
asyncpg
sqlalchemy
passlib[bcrypt]
python-jose
pydantic
email-validator
